Refactor spatial related functions under `spatial` package (#656)
This PR contributes to https://github.com/rapidsai/cuspatial/issues/644, creates a `spatial` package for spatial related functions. Divided into six sub categories: `measure`, `bound`, `index`, `join`, `project`, `filter`.

Also contributes to https://github.com/rapidsai/cuspatial/issues/525, #599

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - H. Thomson Comer (https://github.com/thomcom)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/656
