--- docs/source/api_docs/gis.rst
@@ -1,9 +1,0 @@
-GIS
----
-
-Functions for computing geographic coordinates.
-
-.. currentmodule:: cuspatial
-
-.. autofunction:: cuspatial.haversine_distance
-.. autofunction:: cuspatial.lonlat_to_cartesian
