+++ docs/source/api_docs/index.rst
@@ -9,8 +9,7 @@ This page provides a list of all publicly accessible modules, methods and classe
     :maxdepth: 2
     :caption: API Documentation
 
-    gis
-    spatial_indexing
+    spatial
     trajectory
     geopandas_compatibility
     io
