--- docs/source/api_docs/spatial_indexing.rst
@@ -1,15 +1,0 @@
-Spatial Indexing
-----------------
-
-Spatial indexing functions provide GPU-accelerated point-in-polygon and spatial join operations.
-
-.. currentmodule:: cuspatial
-
-.. autofunction:: cuspatial.quadtree_point_in_polygon
-.. autofunction:: cuspatial.quadtree_point_to_nearest_polyline
-.. autofunction:: cuspatial.point_in_polygon
-.. autofunction:: cuspatial.polygon_bounding_boxes
-.. autofunction:: cuspatial.polyline_bounding_boxes
-.. autofunction:: cuspatial.quadtree_on_points
-.. autofunction:: cuspatial.join_quadtree_and_bounding_boxes
-.. autofunction:: cuspatial.points_in_spatial_window
