+++ python/cuspatial/cuspatial/__init__.py
@@ -1,22 +1,20 @@
 from ._version import get_versions
 from .core import interpolate
-from .core.gis import (
+from .core.spatial import (
     directed_hausdorff_distance,
     haversine_distance,
+    join_quadtree_and_bounding_boxes,
     lonlat_to_cartesian,
-    point_in_polygon,
+    pairwise_linestring_distance,
     polygon_bounding_boxes,
     polyline_bounding_boxes,
-    pairwise_linestring_distance,
-)
-from .core.indexing import quadtree_on_points
-from .core.interpolate import CubicSpline
-from .core.spatial_join import (
-    join_quadtree_and_bounding_boxes,
+    point_in_polygon,
+    points_in_spatial_window,
+    quadtree_on_points,
     quadtree_point_in_polygon,
     quadtree_point_to_nearest_polyline,
 )
-from .core.spatial_window import points_in_spatial_window
+from .core.interpolate import CubicSpline
 from .core.trajectory import (
     derive_trajectories,
     trajectory_bounding_boxes,
