+++ python/cuspatial/cuspatial/core/spatial/filtering.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020, NVIDIA CORPORATION.
+# Copyright (c) 2022, NVIDIA CORPORATION.
 
 from cudf import DataFrame
 from cudf.core.column import as_column
